requirements.txt
customtkinter
mysql-connector-python
Pillow
openpyxl
fpdf
pandas 